# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: lembar API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Baseline contract for planning. The backend repository owns the executable OpenAPI
    document. This file does not authorize implementation of every path at once.
servers:
  - url: /
    description: Current API origin or configured same-origin gateway
...
paths:
...
  /v1/ops/marketing/pages:
    get:
      summary: List marketing pages and authoring state
      tags: [MarketingOps]
      operationId: listMarketingPagesForOps
      responses:
        "200":
          description: Page and version summaries
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/MarketingPageSummary" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/ops/marketing/pages/{slug}:
    get:
      summary: Read marketing page authoring state
      tags: [MarketingOps]
      operationId: getMarketingPageForOps
      parameters:
        - $ref: "#/components/parameters/MarketingSlug"
      responses:
        "200":
          description: Current draft, published version, and version summaries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketingOpsPageResponse" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/ops/marketing/pages/{slug}/draft:
    put:
      summary: Save a structured marketing page draft
      tags: [MarketingOps]
      operationId: saveMarketingPageDraft
      parameters:
        - $ref: "#/components/parameters/MarketingSlug"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaveMarketingDraftRequest" }
      responses:
        "200":
          description: Draft saved with a new revision
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketingOpsPageResponse" }
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/ops/marketing/pages/{slug}/preview:
    get:
      summary: Read authenticated no-store draft preview
      tags: [MarketingOps]
      operationId: previewMarketingPageDraft
      parameters:
        - $ref: "#/components/parameters/MarketingSlug"
      responses:
        "200":
          description: Draft projection for the shared frontend renderer
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketingPageResponse" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/ops/marketing/pages/{slug}/publish:
    post:
      summary: Publish the current validated draft
      tags: [MarketingOps]
      operationId: publishMarketingPageDraft
      parameters:
        - $ref: "#/components/parameters/MarketingSlug"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Immutable version published
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketingOpsPageResponse" }
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/ops/marketing/pages/{slug}/unpublish:
    post:
      summary: Remove the current page from public delivery
      tags: [MarketingOps]
      operationId: unpublishMarketingPage
      parameters:
        - $ref: "#/components/parameters/MarketingSlug"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Public pointer cleared with audit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketingOpsPageResponse" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/ops/marketing/pages/{slug}/versions/{version}/restore:
    post:
      summary: Copy a historical version into a new draft
      tags: [MarketingOps]
      operationId: restoreMarketingPageVersion
      parameters:
        - $ref: "#/components/parameters/MarketingSlug"
        - in: path
          name: version
          required: true
          schema: { type: integer, minimum: 1 }
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "201":
          description: New draft created from the selected version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketingOpsPageResponse" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-lembar_session
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token dengan format "Bearer <token>"
  parameters:
...
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
    IfMatch:
      in: header
      name: If-Match
      required: true
      schema: { type: string }
    MarketingSlug:
      in: path
      name: slug
      required: true
      schema:
        type: string
        enum: [home, untuk-sekolah, harga]
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Authenticated actor lacks the required platform role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found or not visible
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Version, state, or idempotency conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code:
              type: string
              enum:
                - AUTH_REQUIRED
                - PERMISSION_DENIED
                - VALIDATION_FAILED
                - RESOURCE_NOT_FOUND
                - STATE_CONFLICT
                - INTERNAL_ERROR
                - WORKSPACE_ACCESS_DENIED
                - RATE_LIMITED
            message: { type: string }
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
...
    MarketingCta:
      type: object
      additionalProperties: false
      required: [id, label, href, variant, placement, audience, trackingKey, enabled]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        label: { type: string, minLength: 1, maxLength: 120 }
        href: { type: string, minLength: 1, maxLength: 500 }
        variant: { type: string, enum: [primary, secondary, text] }
        placement: { type: string, minLength: 1, maxLength: 80 }
        audience: { type: string, enum: [all, teacher, school] }
        trackingKey: { type: string, minLength: 1, maxLength: 100 }
        enabled: { type: boolean }
        external: { type: boolean, default: false }
        accessibleLabel: { type: [string, "null"], maxLength: 160 }
    MarketingBlockItem:
      type: object
      additionalProperties: false
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string, maxLength: 200 }
        body: { type: [string, "null"], maxLength: 2000 }
        mediaAssetId: { type: [string, "null"] }
        cta: { oneOf: [{ $ref: "#/components/schemas/MarketingCta" }, { type: "null" }] }
    MarketingBlock:
      type: object
      additionalProperties: false
      required: [id, type]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        type:
          type: string
          enum: [hero, product_proof, workflow, audience, trust, pricing, faq, final_cta]
        eyebrow: { type: [string, "null"], maxLength: 120 }
        heading: { type: [string, "null"], maxLength: 300 }
        body: { type: [string, "null"], maxLength: 4000 }
        theme: { type: string, enum: [light, dark, accent], default: light }
        mediaAssetId: { type: [string, "null"] }
        ctas:
          type: array
          maxItems: 4
          items: { $ref: "#/components/schemas/MarketingCta" }
        items:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/MarketingBlockItem" }
    MarketingSeo:
      type: object
      additionalProperties: false
      required: [title, description]
      properties:
        title: { type: string, minLength: 1, maxLength: 70 }
        description: { type: string, minLength: 1, maxLength: 180 }
        imageAssetId: { type: [string, "null"] }
        noIndex: { type: boolean, default: false }
    MarketingPageDocument:
      type: object
      additionalProperties: false
      required: [slug, locale, schemaVersion, version, blocks, seo]
      properties:
        slug: { type: string, enum: [home, untuk-sekolah, harga] }
        locale: { type: string, const: id-ID }
        schemaVersion: { type: integer, minimum: 1 }
        version: { type: integer, minimum: 1 }
        blocks:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: "#/components/schemas/MarketingBlock" }
        seo: { $ref: "#/components/schemas/MarketingSeo" }
    MarketingPageResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/MarketingPageDocument" }
...
    MarketingPageSummary:
      type: object
      required: [slug, locale, state, revision]
      properties:
        slug: { type: string }
        locale: { type: string }
        state: { type: string, enum: [draft, published, unpublished] }
        revision: { type: integer, minimum: 0 }
        publishedVersion: { type: [integer, "null"] }
        updatedAt: { type: string, format: date-time }
    MarketingOpsPageResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [summary, draft, versions]
          properties:
            summary: { $ref: "#/components/schemas/MarketingPageSummary" }
            draft:
              oneOf:
                - $ref: "#/components/schemas/MarketingPageDocument"
                - type: "null"
            versions:
              type: array
              items: { $ref: "#/components/schemas/MarketingPageSummary" }
    SaveMarketingDraftRequest:
      type: object
      additionalProperties: false
      required: [schemaVersion, blocks, seo]
      properties:
        schemaVersion: { type: integer, minimum: 1 }
        blocks:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: "#/components/schemas/MarketingBlock" }
        seo: { $ref: "#/components/schemas/MarketingSeo" }
...
security:
  - BearerAuth: []